requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1

requests-html
pandas
selenium-wire
black
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
twarc